backend/app/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background tasks
celery[redis]==5.3.4
flower==2.0.1
redis==5.0.1

# PDF processing and OCR
PyMuPDF==1.23.9
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0

# AI and embeddings (for future use)
openai==1.3.5
sentence-transformers==2.2.2

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6